requirements.txt
# Core Framework
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0

# Authentication
Flask-Login==0.6.3
bcrypt==4.0.1

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Data Processing
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2

# Configuration
python-dotenv==1.0.0
requests==2.31.0
loguru==0.7.2

# PDF and Barcode Generation
reportlab==4.0.4
python-barcode==0.15.1
Pillow==10.0.0

